feat: add better logic for combined and splitting
When there are multiple sources for a line, and multiple destinations
for a line, then the algorithm doesn't know what to do with it.
Right now, when there are more sources than destinations, we call it a 'combine', when there are equal sources and destinations for a line, we call it a 'move', otherwise it's a 'split'

Documented TODOs of features missing in the tool

Changed the MIN_BLOCK_SIZE from 4 to 1, so that even single line moves
are detected.